epl/requirements.txt
Django
django-bootstrap-v5
django-environ
Pillow
requests
